% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineq1.R
\name{ineq1}
\alias{ineq1}
\title{Inequality for groupwise popn share and income levels}
\usage{
ineq1(y, w)
}
\arguments{
\item{y}{A vector of groupwise income levels}

\item{w}{A vector of groupwise population shares}
}
\value{
A list with four elements:

\item{CV}{Coefficient of variation (scalar)}
\item{Gini}{Gini coefficient (scalar)}
\item{lzdata}{A data frame with cumulative population and income shares}
\item{lzcurve}{A plot of the Lorenz curve (created with ggplots)}
}
\description{
Inequality for groupwise popn share and income levels
}
\examples{
y2011u <- c(700.50,908.92,1118.09,1362.69,1624.86,1887.65,2180.52,2547.94,3062.85,3892.60,5350.06,10281.84)
w2011u <- c(0.05,0.05,rep(0.1,8),0.05,0.05)
res_urb_11 <- ineq1(y=y2011u, w=w2011u)

}
