% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineq3.R
\name{ineq3}
\alias{ineq3}
\title{Inequality for individual income levels}
\usage{
ineq3(y)
}
\arguments{
\item{y}{A vector of individual-level incomes}
}
\value{
A list with the following elements:

\item{CV}{Coefficient of variation (scalar)}
\item{Gini}{Gini coefficient (scalar)}
\item{TH1}{Theil's first measure (scalar)}
\item{lzdata}{A data frame with cumulative population and income shares}
\item{lzcurve}{A plot of the Lorenz curve (created with ggplot2)}
}
\description{
Inequality for individual income levels
}
\examples{
# vector of individual-level income levels
y <- c(100, 200, 300, 400)
myres <- ineq3(y=y)

}
