% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineq4.R
\name{ineq4}
\alias{ineq4}
\title{Inequality for group-level mean income and population levels}
\usage{
ineq4(y, n)
}
\arguments{
\item{y}{A vector of group-level mean income levels (in increasing order)}

\item{n}{A corresponding vector of group-level population levels}
}
\value{
A list with the following elements:

\item{CV}{Coefficient of variation (scalar)}
\item{Gini}{Gini coefficient (scalar)}
\item{TH1}{Theil's first measure (scalar)}
\item{lzdata}{A data frame with cumulative population and income shares}
\item{lzcurve}{A plot of the Lorenz curve (created with ggplot2)}
}
\description{
Inequality for group-level mean income and population levels
}
\examples{
# vector of group-level mean incomes
y <- c(10, 20, 25, 55, 70, 90, 110, 115, 130)
# vector of group-level populations
n <- c(6, 7, 7, 14, 22, 20, 8, 4, 1)
myres <- ineq4(y=y,n=n)

}
