% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineq1.R
\name{ineq1}
\alias{ineq1}
\title{Inequality for groupwise popn share and income levels}
\usage{
ineq1(y, w)
}
\arguments{
\item{y}{A vector of groupwise income levels}

\item{w}{A vector of groupwise population shares}
}
\value{
A list with four elements:

\item{CV}{Coefficient of variation (scalar)}
\item{Gini}{Gini coefficient (scalar)}
\item{lzdata}{A data frame with cumulative population and income shares}
\item{lzcurve}{A plot of the Lorenz curve (created with ggplot2)}
}
\description{
Inequality for groupwise popn share and income levels
}
\examples{
y2011u <- c(700,909,1118,1363,1625,1888,2181,2548,3063,3893,5351,10282)
w2011u <- c(0.05,0.05,rep(0.1,8),0.05,0.05)
res_urb_11 <- ineq1(y=y2011u, w=w2011u)

}
