% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineq4.R
\name{ineq4}
\alias{ineq4}
\title{Inequality for group-level mean income and population levels}
\usage{
ineq4(y, n)
}
\arguments{
\item{y}{A vector of group-level mean income levels}

\item{n}{A vector of group-level population levels}
}
\value{
A list with the following elements:

\item{CV}{Coefficient of variation (scalar)}
\item{Gini}{Gini coefficient (scalar)}
\item{TH1}{Theil's first measure (scalar)}
\item{lzdata}{A data frame with cumulative population and income shares}
\item{lzcurve}{A plot of the Lorenz curve (created with ggplot2)}
}
\description{
Inequality for group-level mean income and population levels
}
\examples{
y <- c(100, 200, 300, 400)
n <- c(50,25,75,25)
myres <- ineq4(y=y,n=n)

}
